feat: logic to stop different token to update/delete other client

+++ server/src/services/client/deleteClient.service.ts
@@ -3,7 +3,7 @@ import { AppError } from "../../errors/appError";
 
 import { Client } from "../../entities/client.entity";
 
-const deleteClientService = async (id: string) => {
+const deleteClientService = async (id: string, clientIdToken: string) => {
   const clientRepository = AppDataSource.getRepository(Client);
 
   const client = await clientRepository.findOneBy({
@@ -12,6 +12,10 @@ const deleteClientService = async (id: string) => {
 
   if (!client) {
     throw new AppError("Client not found", 404);
+  }
+
+  if (clientIdToken !== client.id) {
+    throw new AppError("You don't have permission to do this action", 403);
   }
 
   await clientRepository.delete(id);
